requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
sentence-transformers==2.2.2
faiss-cpu==1.7.4
python-multipart==0.0.6
numpy==1.24.3
requests==2.28.2
pytest==7.3.1
httpx==0.24.0
openai==0.27.6  # For GPT integration
python-dotenv==1.0.0